// types/Colleague.ts
// This interface was generated by http://json2ts.com/ from a sample colleague fetched from the api, and then modified to fit the needs of the application.
export interface Colleague {
    name: string
    email: string
    phoneNumber: string
    office: string
    manager: string
    orgUnit: string
    mainText: string
    gitHub?: string
    twitter?: string
    stackOverflow?: string
    linkedIn?: string
    imagePortraitUrl: string
    imageWallOfLeetUrl: string
    highlighted: boolean
    published: boolean
}